invoice_linking_tool/data_analysis_tools/requirements.txt
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0